+++ api/CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- [DEV-8049] Added `ffi` crate to expose C bindings for wise_units (moved from
+  [WiseUnitsMobile](https://bitbucket.org/agrian/wiseunitsmobile/) with some
+  updates).
+
+### Changed
+
+- Set [#[repr(C)]](https://doc.rust-lang.org/nomicon/other-reprs.html#reprc) 
+  attribute on `Measurement` and related types to safely expose them for FFI.
+
 ## [0.11.0] - 2019-07-17
 
 ### Added